requirements.txt
requests
requests_oauthlib
python-dotenv
MLB-StatsAPI
Pillow
pybaseball